# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "ordinals-minter",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/inscriptions/queue": {
			"post": {
				"operationId": "UploadFileInscription",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Inscriptions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									},
									"fee_option": {
										"type": "string"
									},
									"destination_address": {
										"type": "string"
									}
								},
								"required": [
									"file",
									"fee_option",
									"destination_address"
								]
							}
						}
					}
				}
			}
		},
		"/inscriptions/status": {
			"post": {
				"operationId": "GetInscriptionStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Inscriptions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "depositAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/inscriptions/file": {
			"get": {
				"operationId": "GetInscriptionFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Inscriptions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "depositAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/inscriptions/estimate": {
			"post": {
				"operationId": "GetEstimate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "number",
											"format": "double"
										},
										{
											"items": {},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Inscriptions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"fee_option": {
										"type": "string"
									},
									"bytes_size": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"fee_option",
									"bytes_size"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}